src/LogReader/requirements.txt
falcon==2.0.0
waitress==1.4.4
pymongo[tls]==3.11.0
python-dateutil==2.8.1
hupper==1.10.2
paste==3.5.0
requests==2.24.0
markdown==3.3.3
